+++ libs/web/shell/ui/navbar/src/lib/navbar.service.ts
@@ -24,7 +24,7 @@ export class NavbarService implements OnDestroy {
       .subscribe();
   }
 
-  /** LIFE CYCLE */
+  /** LIFECYCLE */
   public ngOnDestroy(): void {
     this.breadcrumbsSubscription.unsubscribe();
   }