+++ libs/web/shell/ui/sidebar/src/index.ts
@@ -1,3 +1,4 @@
+export * from './lib/sidebar.abstract';
 export * from './lib/sidebar.module';
 export * from './lib/sidebar.component';
 export * from './lib/sidebar.service';